Got 'religion' about correctly using pid_t instead of sloppy 'int's everywhere. The only issue remaining is some of the oplock code that assumes that pid_t's will fit into a 32 bit int. This may need to change... Jeremy.

+++ source/smbwrapper/smbw.c
@@ -97,7 +97,7 @@ void smbw_init(void)
 		DEBUG(2,("SMBW_PREFIX is %s\n", smbw_prefix));
 	}
 
-	slprintf(line,sizeof(line)-1,"PWD_%d", getpid());
+	slprintf(line,sizeof(line)-1,"PWD_%d", (int)getpid());
 	
 	p = smbw_getshared(line);
 	if (!p) {
@@ -1388,7 +1388,7 @@ int smbw_fork(void)
 		smbw_srv_close(srv);
 	}
 
-	slprintf(line,sizeof(line)-1,"PWD_%d", getpid());
+	slprintf(line,sizeof(line)-1,"PWD_%d", (int)getpid());
 	smbw_setshared(line,smbw_cwd);
 
 	/* unblock the parent */
